+++ src/components/Button/Button.story.js
@@ -161,6 +161,15 @@ stories
                 save: 'save',
               }, 'calendar') }>
             { text('customize me!', 'customize me!') }
+          </Button>,
+          <Button
+            key="1"
+            width="default"
+            style="outline"
+            size='small'
+            onClick={ action('I have been clicked') }
+            rightIcon='calendar'>
+            Small icon
           </Button>]
         }
       />